+++ api/quotesRouter.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import fs from "fs";
+
+const router = Router();
+
+
+
+export default router;